requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.7
django-heroku==0.3.1
gunicorn==20.0.4
Jinja2==2.11.3
MarkupSafe==1.1.1
psycopg2==2.8.6
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
tqdm==4.58.0
whitenoise==5.2.0
